Refactor for using Bevy's current master.

+++ src/player.rs
@@ -5,7 +5,7 @@ use crate::world::Velocity;
 pub const WIDTH: f32 = 50.0;
 pub const HEIGHT: f32 = 50.0;
 
-pub const INITIAL_POSITION_X: f32 = -(world::SCREEN_WIDTH as f32 / 2.0) + 128.0;
+pub const INITIAL_POSITION_X: f32 = -world::SCREEN_WIDTH / 2.0 + 128.0;
 pub const INITIAL_POSITION_Y: f32 = HEIGHT / 2.0;
 
 pub const VELOCITY_ON_JUMP: f32 = 600.0;
